embedded_sdmmc

Module fat

FAT16/FAT32 file system implementation

Implements the File Allocation Table file system. Supports FAT16 and FAT32 volumes.

Structs

Bpb

A Boot Parameter Block.

Fat16Info

FAT16 specific data

Fat32Info

FAT32 specific data

FatVolume

Identifies a FAT16 or FAT32 Volume on the disk.

InfoSector

File System Information structure is only present on FAT32 partitions. It may contain a valid number of free clusters and the number of the next free cluster. The information contained in the structure must be considered as advisory only. File system driver implementations are not required to ensure that information within the structure is kept consistent.

OnDiskDirEntry

A 32-byte directory entry as stored on-disk in a directory file.

VolumeName

The name given to a particular FAT formatted volume.

Enums

FatSpecificInfo

Indentifies the supported types of FAT format

FatType

Indentifies the supported types of FAT format

Constants

RESERVED_ENTRIES

Number of entries reserved at the start of a File Allocation Table

Functions

parse_volume

Load the boot parameter block from the start of the given partition and determine if the partition contains a valid FAT16 or FAT32 file system.