embedded_sdmmc::fat

Function parse_volume

pub fn parse_volume<D>(
    block_device: &D,
    lba_start: BlockIdx,
    num_blocks: BlockCount,
) -> Result<VolumeType, Error<D::Error>>
where
    D: BlockDevice,
    D::Error: Debug,

Load the boot parameter block from the start of the given partition and determine if the partition contains a valid FAT16 or FAT32 file system.