embedded_sdmmc

Module filesystem

Generic File System structures

Implements generic file system components. These should be applicable to most (if not all) supported filesystems.

Structs

Attributes

Indicates whether a directory entry is read-only, a directory, a volume label, etc.

ClusterId

Identifies a cluster on disk.

DirEntry

A directory entry, which tells you about other files and directories.

Directory

A handle for an open directory on disk, which closes on drop.

File

A handle for an open file on disk, which closes on drop.

RawDirectory

A handle for an open directory on disk.

RawFile

A handle for an open file on disk.

SearchId

Unique ID used to search for files and directories in the open Volume/File/Directory lists

SearchIdGenerator

A Search ID generator.

ShortFileName

An MS-DOS 8.3 filename. 7-bit ASCII only. All lower-case is converted to upper-case by default.

Timestamp

A Gregorian Calendar date/time, in the local time zone.

Enums

FileError

Errors related to file operations

FilenameError

Various filename related errors that can occur.

Mode

The different ways we can open a file.

Constants

MAX_FILE_SIZE

Maximum file size supported by this library

Traits

TimeSource

Things that impl this can tell you the current time.

ToShortFileName

Describes things we can convert to short 8.3 filenames