embedded_sdmmc
Struct SdMmcSpi

pub struct SdMmcSpi<SPI, CS>
where
    SPI: FullDuplex<u8>,
    CS: OutputPin,
    <SPI as FullDuplex<u8>>::Error: Debug,
{ /* private fields */ }

Represents an SD Card interface built from an SPI peripheral and a Chip Select pin. We need Chip Select to be separate so we can clock out some bytes without Chip Select asserted (which puts the card into SPI mode).

Implementations

impl<SPI, CS> SdMmcSpi<SPI, CS>
where SPI: FullDuplex<u8>, CS: OutputPin, <SPI as FullDuplex<u8>>::Error: Debug,

pub fn new(spi: SPI, cs: CS) -> SdMmcSpi<SPI, CS>

Create a new SD/MMC controller using a raw SPI interface.

pub fn spi(&mut self) -> RefMut<'_, SPI>

Get a temporary borrow on the underlying SPI device. Useful if you need to re-clock the SPI after performing init().

pub fn init(&mut self) -> Result<(), Error>

This routine must be performed with an SPI clock speed of around 100 - 400 kHz. Afterwards you may increase the SPI clock speed.

pub fn deinit(&mut self)

De-init the card so it can’t be used

pub fn card_size_bytes(&self) -> Result<u64, Error>

Return the usable size of this SD card in bytes.

pub fn erase( &mut self, _first_block: BlockIdx, _last_block: BlockIdx, ) -> Result<(), Error>

Erase some blocks on the card.

pub fn erase_single_block_enabled(&self) -> Result<bool, Error>

Can this card erase single blocks?